feat: use global variable manager to manage configuration
