+++ osu.Game.Rulesets.AuthlibInjection/Patches/EndpointPatch.cs
@@ -1,11 +1,5 @@
 #nullable enable
-using System;
-using System.IO;
 using HarmonyLib;
-using Newtonsoft.Json;
-using osu.Framework.Logging;
-using osu.Framework.Platform;
-using osu.Game.Configuration;
 using osu.Game.Online;
 using osu.Game.Rulesets.AuthlibInjection.Configuration;
 
@@ -14,93 +8,9 @@ namespace osu.Game.Rulesets.AuthlibInjection.Patches;
 [HarmonyPatch(typeof(OsuGameBase), nameof(OsuGameBase.CreateEndpoints))]
 public class EndpointPatch
 {
-    private static AuthlibRulesetConfig readFromCommandLine(AuthlibRulesetConfig? config)
+    static void Postfix(ref EndpointConfiguration __result)
     {
-        config ??= new AuthlibRulesetConfig();
-        string[] args = Environment.GetCommandLineArgs();
-        foreach (string arg in args)
-        {
-            string[] split = arg.Split('=');
-
-            string key = split[0];
-            string val = split.Length > 1 ? split[1] : string.Empty;
-            if (string.IsNullOrEmpty(val))
-            {
-                continue;
-            }
-
-            if (!val.StartsWith("http://") && !val.StartsWith("https://"))
-            {
-                val = $"https://{val}";
-            }
-
-            switch (key)
-            {
-                case "--api-url":
-                case "-devserver": // stable like
-                    config.ApiUrl = val;
-                    break;
-                case "--website-url":
-                    config.WebsiteUrl = val;
-                    break;
-                case "--client-id":
-                    config.ClientId = val;
-                    break;
-                case "--client-secret":
-                    config.ClientSecret = val;
-                    break;
-                case "--spectator-url":
-                    config.SpectatorUrl = val;
-                    break;
-                case "--multiplayer-url":
-                    config.MultiplayerUrl = val;
-                    break;
-                case "--metadata-url":
-                    config.MetadataUrl = val;
-                    break;
-                case "--bss-url":
-                    config.BeatmapSubmissionServiceUrl = val;
-                    break;
-            }
-        }
-
-        return config;
-    }
-
-    static void Postfix(OsuGameBase __instance, ref EndpointConfiguration __result)
-    {
-        // try get game folder
-        var localConfig = Traverse.Create(__instance).Property("LocalConfig").GetValue<OsuConfigManager>();
-        string configPath = AuthlibRulesetConfig.CONFIG_FILE_NAME;
-        if (localConfig != null)
-        {
-            var storage = Traverse.Create(localConfig).Field("storage").GetValue<Storage>();
-            configPath = storage.GetFullPath(AuthlibRulesetConfig.CONFIG_FILE_NAME);
-        }
-
-        if (File.Exists(configPath) == false)
-        {
-            return;
-        }
-
-        string config = File.ReadAllText(configPath);
-        if (string.IsNullOrEmpty(config))
-        {
-            return;
-        }
-
-        AuthlibRulesetConfig? authlibLocalConfig = null;
-        try
-        {
-            authlibLocalConfig = JsonConvert.DeserializeObject<AuthlibRulesetConfig>(config);
-        }
-        catch (JsonException)
-        {
-            Logger.Log("[AuthlibInjection] Failed to parse authlib_local_config.json, please check the json format.",
-                level: LogLevel.Error);
-        }
-
-        authlibLocalConfig = readFromCommandLine(authlibLocalConfig);
+        var authlibLocalConfig = GlobalConfigManager.Instance;
 
         if (!string.IsNullOrEmpty(authlibLocalConfig.ApiUrl))
         {
